refactor(app): extract useExternalFileChanges hook (5tp)
Moves externallyChangedFiles, filesWithDiskConflict, the file-watcher IPC
subscription effect, and handleKeepCurrentFile out of App.tsx into
src/hooks/useExternalFileChanges.ts.
